Move task manager to core
